Remove pandas_add_column
