+++ python/ray/data/tests/test_map.py
@@ -473,6 +473,25 @@ def test_add_column(ray_start_regular_shared):
         ray.data.range(5).add_column("foo", lambda x: x["id"] + 1, batch_format="foo")
 
 
+def test_add_column_to_pandas(ray_start_regular_shared):
+    # Refer to issue https://github.com/ray-project/ray/issues/51758
+    ds = ray.data.from_pandas(
+        pd.DataFrame({"a": list(range(20))}), override_num_blocks=2
+    )
+
+    ds = ds.add_column(
+        "foo1", lambda df: pd.Series([1] * len(df)), batch_format="pandas"
+    )
+    ds = ds.add_column(
+        "foo2", lambda df: pd.DatetimeIndex([1] * len(df)), batch_format="pandas"
+    )
+    ds = ds.add_column(
+        "foo3", lambda df: pd.DataFrame({"foo": [1] * len(df)}), batch_format="pandas"
+    )
+    for row in ds.iter_rows():
+        assert row["foo1"] == 1 and row["foo2"] == pd.Timestamp(1) and row["foo3"] == 1
+
+
 @pytest.mark.parametrize(
     "names, expected_schema",
     [
